refactor(integrations): Remove an unnecessary node package from the integrations
