+++ integration/docker/cucumber/environment.js
@@ -1,5 +1,6 @@
-module.exports = {
+const sharedEnv = require('../../../lib/environment');
+
+module.exports = Object.assign({}, sharedEnv, {
   username: process.env.DOCKER_USERNAME,
   password: process.env.DOCKER_PASSWORD,
-  longTimeout: process.env.WEBDRIVER_LONG_TIMEOUT,
-};
+});
