+++ integration/selenium/cucumber/steps.js
@@ -1,11 +1,13 @@
+const sharedEnv = require('../../../lib/environment');
+
 module.exports = function () {
   const autUrl = process.env.AUT_SERVER_URL || '.';
 
-  this.Given(/^I have loaded the web application$/, { timeout: 10000 }, async function () {
+  this.Given(/^I have loaded the web application$/, { timeout: sharedEnv.timeout }, async function () {
     return this.driver.get(`${autUrl}`);
   });
 
-  this.Given(/^I have loaded the web application with "([^"]+)"$/, { timeout: 10000 }, async function (url) {
+  this.Given(/^I have loaded the web application with "([^"]+)"$/, { timeout: sharedEnv.timeout }, async function (url) {
     return this.driver.get(url);
   });
 };
