+++ integration/snyk.io/cucumber/environment.js
@@ -1,7 +1,6 @@
-const env = require('../../../lib/environment');
-const merge = require('lodash.merge');
+const sharedEnv = require('../../../lib/environment');
 
-module.exports = merge(env, {
+module.exports = Object.assign({}, sharedEnv, {
   snykUrl: process.env.SNYK_URL || 'http://snyk:8086/',
   snykFailureLevel: process.env.SNYK_FAILURE_LEVEL || 'high',
 });
