+++ integration/sslyze/.eslintrc
@@ -1,0 +1,7 @@
+{
+    "rules": {
+        "import/no-extraneous-dependencies": "off",
+        "import/no-unresolved": "off",
+        "func-names": "off"
+    }
+}
