+++ integration/sslyze/cucumber/environment.js
@@ -1,7 +1,6 @@
-const env = require('../../../lib/environment');
-const merge = require('lodash.merge');
+const sharedEnv = require('../../../lib/environment');
 
-module.exports = merge(env, {
+module.exports = Object.assign({}, sharedEnv, {
   serverHostName: process.env.AUT_SERVER_URL || '.',
   sslyzeServerUrl: process.env.SSLYZE_SERVER_URL || 'http://sslyze:8081/',
 });
